bookshelf.js -> lib/index.js
